changed variable data format

+++ EODHistoricalData.Wrapper/Model/BulkFundamental/Highlights.cs
@@ -10,7 +10,7 @@ namespace EOD.Model.BulkFundamental
         /// <summary>
         /// 
         /// </summary>
-        public long? MarketCapitalization { get; set; }
+        public double? MarketCapitalization { get; set; }
         /// <summary>
         /// 
         /// </summary>
