+++ EODHistoricalData.Wrapper/Model/Fundamental/Highlights.cs
@@ -12,7 +12,7 @@ namespace EOD.Model.Fundamental
         /// <summary>
         /// 
         /// </summary>
-        public long? MarketCapitalization { get; set; }
+        public double? MarketCapitalization { get; set; }
         /// <summary>
         /// 
         /// </summary>
